requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pystray>=0.19.0
mss>=9.0.0
